deleting the moved files
