--- components/LandingStudentRightsGuide.tsx
@@ -1,204 +1,0 @@
-import React from 'react'
-import Divider from '@mui/material/Divider'
-
-function LandingStudentRights() {
-  return (
-    <div>
-      <div
-        style={{
-          width: '90%',
-          marginLeft: '20px',
-          marginTop: '30px',
-          marginBottom: '30px',
-        }}
-      >
-        <h1>Student Rights Guide</h1>
-        <p>Last updated January 2021</p>
-        <Divider />
-        <br />
-        <p>
-          Our PRS tool can help you file a PRS complaint by walking you through
-          the process of writing and submitting the complaint. The tool asks
-          questions about your concerns to identify specific legal rights that
-          may have been violated.
-        </p>
-        <br />
-        <p>
-          These guides provide an overview of the rights that are included in
-          the guides, so that you can get an idea of whether your concerns fit
-          within the information included in the guides. Explore the guides
-          below to learn more about what kinds of situations are included in the
-          tool.
-        </p>
-        <br />
-        <p> - </p>
-        <br />
-        <h2 id="school-discipline">School Discipline & Education</h2>
-        <br />
-        <p>
-          Students have the right to be free from unjust discipline. The tool
-          covers students’ legal rights when a school is trying to suspend or
-          expel them.
-        </p>
-        <br />
-        <p>
-          <em>
-            <strong>What is a suspension?</strong>
-          </em>
-        </p>
-        <br />
-        <p>
-          A student is suspended if they are excluded from their classroom. This
-          includes both in-school and out-of-school suspensions.
-        </p>
-        <br />
-        <p>A student is suspended if: </p>
-        <br />
-        <ul style={{ marginLeft: '20px' }}>
-          <li>They are not allowed to come to school</li>
-          <li>They are sent home early</li>
-          <li>
-            They are in the school building, but they aren&apos;t allowed to be
-            in their regular classroom for more than half of the school day. For
-            example, if they have to sit in the principla&apos;s office.
-          </li>
-        </ul>
-        <br />
-        <p>
-          <em>
-            <strong>What is an expulsion?</strong>
-          </em>
-        </p>
-        <br />
-        <p>
-          A student is expelled if they are excluded from their classroom for
-          more than 90 days. This includes students who are excluded
-          indefinitely or permanently.
-        </p>
-        <br />
-        <p>
-          <strong>
-            There are different laws and requirements depending on the reason
-            the school is disciplining the student.{' '}
-          </strong>
-        </p>
-        <br />
-        <p>
-          The tool covers the requirements for all three types of discipline:{' '}
-        </p>
-        <br />
-        <p>
-          <strong>Code of Conduct Violations:</strong>
-          <br />
-          The school is accusing the student of breaking a school rule.
-        </p>
-        <br />
-        <p>
-          <strong>Expellable Offenses: </strong>
-          <br />
-          The school is accusing the student of bringing a weapon or drugs to
-          school, or assaulting school staff.
-        </p>
-        <br />
-        <p>
-          <strong>Felony Complaint:</strong>
-          <br />
-          The school is excluding the student because of a felony complaint or
-          adjudication that is unrelated to school.
-        </p>
-        <br />
-        <p>
-          <em>
-            <strong>
-              How do I know why the school is disciplining the student?
-            </strong>
-          </em>
-        </p>
-        <br />
-        <p>
-          The school has to tell the student and their parent/guardian why
-          they’re disciplining the student. If the school told you about the
-          discipline in writing, the information should be in that letter or
-          email. Call the EdLaw Project intake line if you’re not sure why the
-          student is being disciplined.
-        </p>
-        <br /> <p> - </p> <br />
-        <h2 id="special-education">Special Education</h2>
-        <br />
-        <p>
-          Students with disabilities have the right to a public education that
-          meets their needs. The tool covers the requirements for the process of
-          identifying whether a student has a disability, determining what kinds
-          of services they need, and making a plan for the student to access the
-          services.
-        </p>
-        <br />
-        <p>
-          <strong>
-            Please note that the tool only covers the requirements for the
-            special education process.
-          </strong>
-        </p>
-        <br />
-        <p>
-          The questions will identify issues where the school clearly did or
-          didn’t do something (for example, the school didn’t follow the right
-          timeline for an evaluation, or the school isn’t providing the services
-          in the student’s IEP).
-        </p>
-        <br />
-        <p>
-          The questions don’t cover disagreements about what the student needs
-          (for example, whether the student is eligible for special education
-          services, or whether the student needs to be in a smaller classroom),
-          because those situations require more specific details about the
-          student.
-        </p>
-        <br />
-        <p>
-          Sometimes, a school follows the right process, but the special
-          education plan the school offers doesn’t work for the student. If that
-          is the case, the questions in the tool won’t address all of your
-          concerns. Call the EdLaw Project intake line if you have questions
-          about whether your situation can be addressed through a PRS complaint,
-          or if other steps might make more sense.
-        </p>
-        <br /> <p> - </p> <br />
-        <h2 id="bullying">Bullying</h2>
-        <br />
-        <p>
-          A student is being bullied when another student’s (or group of
-          students’) actions hurt them. Bullying can be when: <br /> <br />
-        </p>
-        <ul style={{ marginLeft: '20px' }}>
-          <li>
-            A student harms another student, or makes the student think that
-            they’re going to harm them. The harm can be physical or emotional.
-          </li>
-          <li>
-            A student damages another student’s property, or makes the student
-            think they’re going to damage their property.
-          </li>
-          <li>
-            A student creates a hostile environment at school that interferes
-            with another student’s education.
-          </li>
-          <li>
-            A student interferes with another student’s right to learn at school{' '}
-          </li>
-          <li>A student disrupts the school’s operations</li>
-        </ul>
-        <br />
-        <p>
-          Students have the right to have their schools prevent and respond to
-          bullying. The tool covers the requirements the school has to follow
-          when they learn a student has been bullied, and the requirements for
-          the school to have a Bullying Prevention Plan in place.
-        </p>
-        <br />
-      </div>
-    </div>
-  )
-}
-
-export default LandingStudentRights
