+++ pages/formDev/formik.tsx
@@ -1,8 +1,8 @@
 import React, { useContext } from 'react'
 import * as Yup from 'yup'
-import { MyTextInput } from '../../components/MyInput'
-import { MySelect } from '../../components/MySelect'
-import { MyCheckbox } from '../../components/MyCheckbox'
+import { MyTextInput } from '../../components/DynamicForm/MyInput'
+import { MySelect } from '../../components/FormikExample/MySelect'
+import { MyCheckbox } from '../../components/FormikExample/MyCheckbox'
 import { useRouter } from 'next/router'
 import { FormCtx, FormValues } from '../../utils/FormContext'
 import { Form, Formik } from 'formik'
